requirements.txt
telethon
python-dotenv
pandas
tqdm
loguru
